+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/ybiquitous/aufgaben/compare/0.8.4...HEAD)
 
+- Check committed files on `release` task [#49](https://github.com/ybiquitous/aufgaben/pull/49)
+
 ## 0.8.4
 
 [Full diff](https://github.com/ybiquitous/aufgaben/compare/0.8.3...0.8.4)